fixup! Gut field tests
