+++ test/celeritas/field/Steppers.test.cc
@@ -8,6 +8,9 @@
 
 #include "Steppers.test.hh"
 
+#include "celeritas_config.h"
+#include "corecel/Assert.hh"
+#include "corecel/Macros.hh"
 #include "corecel/Types.hh"
 #include "corecel/cont/Range.hh"
 #include "celeritas/Constants.hh"
@@ -154,39 +157,3 @@ TEST_F(SteppersTest, host_dormand_prince_547)
     // Test the Dormand-Prince 547(M) stepper
     this->run_stepper<UniformField, DormandPrinceStepper>(field);
 }
-
-//---------------------------------------------------------------------------//
-// DEVICE TESTS
-//---------------------------------------------------------------------------//
-TEST_F(SteppersTest, TEST_IF_CELER_DEVICE(device_helix))
-{
-    // Run the ZHelix kernel
-    param.nstates = 32 * 512;
-    auto output   = helix_test(param);
-
-    // Check stepper results
-    check_result(output);
-}
-
-//---------------------------------------------------------------------------//
-TEST_F(SteppersTest, TEST_IF_CELER_DEVICE(device_classical_rk4))
-{
-    // Run the classical Runge-Kutta kernel
-    param.nstates = 32 * 512;
-    auto output   = rk4_test(param);
-
-    // Check stepper results
-    check_result(output);
-}
-
-//---------------------------------------------------------------------------//
-TEST_F(SteppersTest, TEST_IF_CELER_DEVICE(device_dormand_prince_547))
-{
-    // Run the Dormand-Prince 547(M) kernel
-    param.nstates = 32 * 512;
-    auto output   = dp547_test(param);
-
-    // Check stepper results
-    check_result(output);
-}
-//---------------------------------------------------------------------------//
